VPSSessionProtocol

See merge request arvrlab/vps-client-apps/public/vps-unity3d-sdk!27

+++ Assets/Scripts/GPS/GPSData.cs
@@ -13,7 +13,7 @@ namespace ARVRLab.VPSService
         public double Latitude;
         public double Longitude;
         public double Altitude;
-        public float Accuracy;
+        public double Accuracy;
         public double Timestamp;
 
         public GPSData()
